eks: x100-Operator v2
Redesigned the operator to allow for flexibility
to handle scenarios such as reset/reboot and allow
for pod sequencing which helps prevent critical
race conditions between hwmgr and firmware pod.
Design also adds support for termination sequence
to allow for a graceful shutdown of resources.

Changes the state machine to remove rollback handling
as a state and fixes the rollback path to execute properly.

Adds robust labeling scheme to allow for better state
management in case of unexpected errors.

Change-Id: I3d2730cc5533a0377615b1b2a904fe805a70c87f