Counter NFD fluctuation and label propagation delays
Change allows for waiting on nfd fluctuations to settle down.
Operator allows for deferring the processing of node to check
for presence of x100 card which relies on pci label attached
by nfd resource.

We also provide a mechanism to force the updates to propagate to
server before we flag labeling as successful.

Change-Id: Ia7d321ce3c4f4a7ceb72c82164df0df1810535aa