Handle empty policy lists in runtime
Change adds the ability to identify and handle
the policy whose node selector list gets empty
in runtime due to movement of nodes to other
policies. This also adds a safeguard from application
of such policies to x100 nodes that are with out any
policies in the cluster at any given point.

Change-Id: I0a7e93fce2b551e179225bd4702f5059f2363b4f